+++ examples/src/test/java/org/chocosolver/examples/integer/MagicSquareTest.java
@@ -9,7 +9,7 @@
  */
 package org.chocosolver.examples.integer;
 
-import org.chocosolver.pf4cs.SetUpException;
+import org.chocosolver.parser.SetUpException;
 import org.chocosolver.solver.Cause;
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.exception.ContradictionException;