+++ parsers/src/main/java/org/chocosolver/parser/RegParser.java
@@ -10,7 +10,6 @@
 package org.chocosolver.parser;
 
 import gnu.trove.set.hash.THashSet;
-import org.chocosolver.pf4cs.SetUpException;
 import org.chocosolver.solver.*;
 import org.chocosolver.solver.learn.XParameters;
 import org.chocosolver.solver.search.loop.move.MoveBinaryDFS;
@@ -36,7 +35,7 @@ import java.util.List;
  * Project: choco-parsers.
  */
 public abstract class RegParser implements IParser {
-
+    public static boolean PRINT_LOG = true;
     /**
      * Name of the parser
      */
@@ -171,12 +170,15 @@ public abstract class RegParser implements IParser {
         return defaultSettings;
     }
 
-    @Override
+    /**
+     * Create the solver
+     */
+    public abstract void createSolver();
+
     public final void addListener(ParserListener listener) {
         listeners.add(listener);
     }
 
-    @Override
     public final void removeListener(ParserListener listener) {
         listeners.remove(listener);
     }
@@ -184,7 +186,7 @@ public abstract class RegParser implements IParser {
     @Override
     public final boolean setUp(String... args) throws SetUpException {
         listeners.forEach(ParserListener::beforeParsingParameters);
-        System.out.printf("%s %s\n", getCommentChar(), Arrays.toString(args));
+        if(PRINT_LOG)System.out.printf("%s %s\n", getCommentChar(), Arrays.toString(args));
         CmdLineParser cmdparser = new CmdLineParser(this);
         try {
             cmdparser.parseArgument(args);
@@ -241,12 +243,12 @@ public abstract class RegParser implements IParser {
         Solver solver = portfolio.getModels().get(0).getSolver();
         if (nb_cores == 1) {
             if (exp) {
-                System.out.printf("%s exp is on\n", getCommentChar());
+                if(PRINT_LOG)System.out.printf("%s exp is on\n", getCommentChar());
                 solver.setLearningSignedClauses();
                 // THEN PARAMETERS
                 XParameters.DEFAULT_X = dftexp;
-                XParameters.PROOF = XParameters.FINE_PROOF = false;
-                XParameters.PRINT_CLAUSE = false;
+                XParameters.PROOF = XParameters.FINE_PROOF = true;
+                XParameters.PRINT_CLAUSE = true;
                 XParameters.ASSERT_UNIT_PROP = true; // todo : attention aux clauses globales
                 XParameters.ASSERT_NO_LEFT_BRANCH = false;
                 XParameters.INTERVAL_TREE = true;
@@ -255,18 +257,18 @@ public abstract class RegParser implements IParser {
                 }
             }
             if (free) {
-                System.out.printf("%s set search to: (%s,%s) + %s\n", getCommentChar(), varH, valH, restarts.pol);
+                if(PRINT_LOG)System.out.printf("%s set search to: (%s,%s) + %s\n", getCommentChar(), varH, valH, restarts.pol);
                 if(lc > 0 || cos || last){
-                    System.out.printf("%s add techniques: ", getCommentChar());
+                    if(PRINT_LOG)System.out.printf("%s add techniques: ", getCommentChar());
                     if(cos){
-                        System.out.print("-cos ");
+                        if(PRINT_LOG)System.out.print("-cos ");
                     }else if(lc>0){
-                        System.out.printf("-lc %d ", lc);
+                        if(PRINT_LOG)System.out.printf("-lc %d ", lc);
                     }
                     if(last){
-                        System.out.print("-last");
+                        if(PRINT_LOG)System.out.print("-last");
                     }
-                    System.out.print("\n");
+                    if(PRINT_LOG)System.out.print("\n");
                 }
                 IntVar obj = (IntVar) solver.getObjectiveManager().getObjective();
                 IntVar[] dvars = Arrays.stream(solver.getMove().getStrategy().getVariables())
@@ -292,6 +294,8 @@ public abstract class RegParser implements IParser {
                     solver.setSearch(Search.lastConflict(solver.getSearch(), lc));
                 }
                 restarts.declare(solver);
+                //solver.plugMonitor(new BackjumpRestart(solver));
+                //solver.showDecisions();
             }
         }
         for (int i = 0; i < nb_cores; i++) {