+++ solver/src/main/java/org/chocosolver/solver/Solver.java
@@ -44,7 +44,6 @@ import org.chocosolver.util.criteria.Criterion;
 
 import java.io.PrintStream;
 import java.util.*;
-import java.util.function.Predicate;
 
 import static org.chocosolver.solver.Solver.Action.*;
 import static org.chocosolver.solver.constraints.Constraint.Status.FREE;
@@ -329,12 +328,9 @@ public class Solver implements ISolver, IMeasures, IOutputFactory {
             //noinspection unchecked
             Set<Constraint> instances = (Set<Constraint>) mModel.getHook("cinstances");
             if (instances != null) {
-                Predicate<Constraint> freed =
-                        c -> (c.getStatus() == FREE) || (
-                                c.getOpposite() != null && c.getOpposite().getStatus() == FREE);
                 Optional<Constraint> undeclared = instances
                         .stream()
-                        .filter(freed)
+                        .filter(c -> (c.getStatus() == FREE))
                         .findFirst();
                 if (undeclared.isPresent()) {
                     getErr().print(
@@ -1348,6 +1344,6 @@ public class Solver implements ISolver, IMeasures, IOutputFactory {
 
     @Override
     public PrintStream getErr() {
-        return out;
+        return err;
     }
 }