+++ solver/src/main/java/org/chocosolver/solver/search/IResolutionHelper.java
@@ -39,6 +39,7 @@ import java.util.stream.StreamSupport;
  * @author Jean-Guillaum Fages
  * @author Charles Prud'homme
  * @author Guillaume Lelouet
+ * @author Dimitri Justeau-Allaire (dimitri.justeau@gmail.com)
  * @since 25/04/2016.
  */
 public interface IResolutionHelper extends ISelf<Solver> {
@@ -328,9 +329,12 @@ public interface IResolutionHelper extends ISelf<Solver> {
             int opt = ref().getObjectiveManager().getBestSolutionValue().intValue();
             ref().reset();
             ref().getModel().clearObjective();
-            ref().getModel().arithm(objective, "=", opt).post();
-			if(defaultS) ref().setSearch(Search.defaultSearch(ref().getModel()));// best bound (in default) is only for optim
-            return findAllSolutions(stop);
+            Constraint forceOptimal = ref().getModel().arithm(objective, "=", opt);
+            forceOptimal.post();
+            if(defaultS) ref().setSearch(Search.defaultSearch(ref().getModel()));// best bound (in default) is only for optim
+            List<Solution> solutions = findAllSolutions(stop);
+            ref().getModel().unpost(forceOptimal);
+            return solutions;
         } else {
             ref().removeStopCriterion(stop);
             return Collections.emptyList();
@@ -393,9 +397,40 @@ public interface IResolutionHelper extends ISelf<Solver> {
             int opt = ref().getObjectiveManager().getBestSolutionValue().intValue();
             ref().reset();
             ref().getModel().clearObjective();
-            ref().getModel().arithm(objective, "=", opt).post();
+            Constraint forceOptimal = ref().getModel().arithm(objective, "=", opt);
+            forceOptimal.post();
+            ref().getModel().getEnvironment().save(() -> ref().getModel().unpost(forceOptimal));
             if(defaultS) ref().setSearch(Search.defaultSearch(ref().getModel()));// best bound (in default) is only for optim
-            return streamSolutions(stop);
+            Spliterator<Solution> it = new Spliterator<Solution>() {
+
+                @Override
+                public boolean tryAdvance(Consumer<? super Solution> action) {
+                    if (ref().solve()) {
+                        action.accept(new Solution(ref().getModel()).record());
+                        return true;
+                    }
+                    ref().getModel().unpost(forceOptimal);
+                    ref().removeStopCriterion(stop);
+                    return false;
+                }
+
+                @Override
+                public Spliterator<Solution> trySplit() {
+                    return null;
+                }
+
+                @Override
+                public long estimateSize() {
+                    return Long.MAX_VALUE;
+                }
+
+                @Override
+                public int characteristics() {
+                    return Spliterator.ORDERED | Spliterator.DISTINCT | Spliterator.NONNULL | Spliterator.CONCURRENT;
+                }
+
+            };
+            return StreamSupport.stream(it, false);
         } else {
             ref().removeStopCriterion(stop);
             return Stream.empty();